+++ sakura_core/doc/CDocOutline.cpp
@@ -86,7 +86,7 @@ int CDocOutline::ReadRuleFile( const WCHAR* pszFilename, SOneRule* pcOneRule, in
 		strLine = file.ReadLineW();
 		pszWork = wcsstr( strLine.c_str(), pszDelimit );
 		if( nullptr != pszWork && 0 < strLine.length() && strLine[0] != cComment ){
-			int nLen = pszWork - strLine.c_str();
+			int nLen = int(pszWork - strLine.c_str());
 			if( nLen < LINEREADBUFSIZE ){
 				// szLine == 「key1,key2」
 				wmemcpy(szLine, strLine.c_str(), nLen);
@@ -120,7 +120,7 @@ int CDocOutline::ReadRuleFile( const WCHAR* pszFilename, SOneRule* pcOneRule, in
 					if( nullptr != pszGroupDel && 0 < pszWork[0] != L'\0' ){
 						// pszWork = 「titleRep /// group」
 						// pszGroupDel = 「 /// group」
-						int nTitleLen = pszGroupDel - pszWork; // Len == 0 OK
+						int nTitleLen = int(pszGroupDel - pszWork); // Len == 0 OK
 						if( nTitleLen < int(std::size(szText)) ){
 							wcsncpy_s(szText, std::size(szText), pszWork, nTitleLen);
 						}else{
@@ -150,7 +150,7 @@ int CDocOutline::ReadRuleFile( const WCHAR* pszFilename, SOneRule* pcOneRule, in
 				pcOneRule[nCount].szMatch[255] = L'\0';
 				pcOneRule[nCount].szGroupName[255] = L'\0';
 				pcOneRule[nCount].nLv = nLv;
-				pcOneRule[nCount].nLength = wcslen(pcOneRule[nCount].szMatch);
+				pcOneRule[nCount].nLength = (int)wcslen(pcOneRule[nCount].szMatch);
 				pcOneRule[nCount].nRegexOption = regexOption;
 				pcOneRule[nCount].nRegexMode = bRegexRep2 ? 1 : 0; // 文字列が正しい時だけReplaceMode
 				nCount++;
@@ -274,13 +274,13 @@ void CDocOutline::MakeFuncList_RuleFile( CFuncInfoArr* pcFuncInfoArr, std::wstri
 	if( test[0].nLength == 0 ){
 		const wchar_t* g = test[0].szGroupName;
 		wcscpy(pszStack[0], g);
-		nLvStack[0] = test[0].nLv;
+		nLvStack[0] = int(test[0].nLv);
 		const wchar_t *p = wcschr(g, L',');
 		int len;
 		if( p != nullptr ){
-			len = p - g;
+			len = int(p - g);
 		}else{
-			len = wcslen(g);
+			len = (int)wcslen(g);
 		}
 		CNativeW mem;
 		mem.SetString(g, len);
